Restructure painting order file
